+++ sportsdataverse/nhl/__init__.py
@@ -50,6 +50,7 @@ from sportsdataverse.nhl.nhl_game_rosters import *
 from sportsdataverse.nhl.nhl_loaders import *
 from sportsdataverse.nhl.nhl_pbp import *
 from sportsdataverse.nhl.nhl_records import *
+from sportsdataverse.nhl.nhl_records_extra import *
 from sportsdataverse.nhl.nhl_schedule import *
 from sportsdataverse.nhl.nhl_teams import *
 from sportsdataverse.nhl.nhl_espn_ext import *